fix: update code style
