+++ spring-cloud-alibaba-dependencies/pom.xml
@@ -214,11 +214,6 @@
                 <artifactId>dubbo-remoting-netty4</artifactId>
                 <version>${dubbo.version}</version>
             </dependency>
-            <dependency>
-                <groupId>org.apache.dubbo</groupId>
-                <artifactId>dubbo-serialization-hessian2</artifactId>
-                <version>${dubbo.version}</version>
-            </dependency>
 
             <!-- Own dependencies  -->
             <dependency>
