+++ spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-api/src/main/java/com/alibaba/api/ProviderService.java
@@ -14,22 +14,31 @@
  * limitations under the License.
  */
 
-package com.alibaba.dubbo.consumer.api;
+package com.alibaba.api;
 
-import com.alibaba.dubbo.consumer.entity.User;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 
 /**
  * @author :Lictory
- * @date : 2024/08/01
  */
 @FeignClient("provider")
-public interface FooService {
-    @GetMapping("/api/test")
-    String foo();
+public interface ProviderService {
+	/**
+	 * string test.
+	 *
+	 * @return string
+	 */
+	@GetMapping("/api/get")
+	String hi();
 
-    @GetMapping("/user")
-    User getUser();
+	/**
+	 * post test.
+	 *
+	 * @param userId id
+	 */
+	@PostMapping("/api/post/{userId}")
+	void post(@PathVariable("userId") Integer userId);
 }
-
