--- spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-api/src/main/java/com/alibaba/dubbo/consumer/entity/User.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2013-2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.dubbo.consumer.entity;
-
-import java.io.Serializable;
-
-/**
- * @author :Lictory
- * @date : 2024/08/01
- */
-public class User implements Serializable {
-    private int id=1;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-}
