+++ spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-consumer/src/main/java/com/alibaba/dubbo/consumer/service/TestService.java
@@ -16,7 +16,9 @@
 
 package com.alibaba.dubbo.consumer.service;
 
-import com.alibaba.dubbo.consumer.api.ProviderService;
+
+import com.alibaba.api.ProviderService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -29,18 +31,18 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 public class TestService {
-    @Autowired
-    private ProviderService providerService;
+	@Autowired
+	private ProviderService providerService;
 
-    @GetMapping("/test/get")
-    public String testGet() {
-        return providerService.hi();
-    }
+	@GetMapping("/test/get")
+	public String testGet() {
+		return providerService.hi();
+	}
 
-    @PostMapping("/test/post/{userId}")
-    public void testPost(@PathVariable("userId") Integer userId) {
-        providerService.post(userId);
-    }
+	@PostMapping("/test/post/{userId}")
+	public void testPost(@PathVariable("userId") Integer userId) {
+		providerService.post(userId);
+	}
 }
