+++ spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-provider/src/main/java/com/alibaba/dubbo/provider/ProviderApplication.java
@@ -28,7 +28,7 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 @SpringBootApplication
 @EnableFeignClients
 public class ProviderApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(ProviderApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(ProviderApplication.class, args);
+	}
 }
