+++ spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-provider/src/main/java/com/alibaba/dubbo/provider/service/impl/ProviderServiceImpl.java
@@ -17,9 +17,8 @@
 package com.alibaba.dubbo.provider.service.impl;
 
 
-import com.alibaba.dubbo.consumer.api.ProviderService;
-import com.alibaba.dubbo.consumer.entity.User;
-import org.apache.dubbo.config.annotation.DubboService;
+import com.alibaba.api.ProviderService;
+
 import org.springframework.stereotype.Service;
 
 /**
@@ -28,14 +27,14 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class ProviderServiceImpl implements ProviderService {
-    @Override
-    public String hi() {
-        return "Hi Spring Cloud Alibaba 测试成功";
-    }
+	@Override
+	public String hi() {
+		return "Hi Spring Cloud Alibaba 测试成功";
+	}
 
-    @Override
-    public void post(Integer userId) {
-        System.out.println("sca post Test 测试"+userId);
-    }
+	@Override
+	public void post(Integer userId) {
+		System.out.println("sca post Test 测试" + userId);
+	}
 
 }
